[Step5] refactor: Step4 피드백 반영
