+++ src/main/kotlin/racingcar/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.util
+package racingcar.domain
 
 fun interface NumberGenerator {
     fun generateNumber(): Int
