--- src/main/kotlin/racingcar/util/CarMoveNumberGenerator.kt
@@ -1,7 +1,0 @@
-package racingcar.util
-
-class CarMoveNumberGenerator : NumberGenerator {
-    override fun generateNumber(): Int {
-        return 4
-    }
-}
