--- src/main/kotlin/racingcar/util/CarStopNumberGenerator.kt
@@ -1,7 +1,0 @@
-package racingcar.util
-
-class CarStopNumberGenerator : NumberGenerator {
-    override fun generateNumber(): Int {
-        return 3
-    }
-}
