--- src/main/kotlin/racingcar/util/RandomGenerator.kt
@@ -1,9 +1,0 @@
-package racingcar.util
-
-import kotlin.random.Random
-
-class RandomGenerator : NumberGenerator {
-    override fun generateNumber(): Int {
-        return Random.nextInt(0, 10)
-    }
-}
