+++ src/test/kotlin/racingcar/RacingCarGameTest.kt
@@ -6,15 +6,13 @@ import io.kotest.data.row
 import io.kotest.matchers.shouldBe
 import racingcar.domain.RacingCar
 import racingcar.domain.RacingCarGame
-import racingcar.util.CarMoveNumberGenerator
-import racingcar.util.CarStopNumberGenerator
 
 class RacingCarGameTest : StringSpec({
 
     val racingCarGame = RacingCarGame(listOf(RacingCar("test")))
 
     "1 ~ 9 까지의 입력 중 4 미만일 경우 차가 전진할 수 없다." {
-        racingCarGame.racing(CarStopNumberGenerator())
+        racingCarGame.racing { 0 }
         racingCarGame.racingCars[0].position shouldBe 0
     }
 
@@ -27,7 +25,7 @@ class RacingCarGameTest : StringSpec({
             row(5),
             row(6),
         ) { position ->
-            racingCarGame.racing(CarMoveNumberGenerator())
+            racingCarGame.racing { 4 }
             racingCarGame.racingCars[0].position shouldBe position
         }
     }
@@ -37,7 +35,7 @@ class RacingCarGameTest : StringSpec({
             listOf(RacingCar("1등"), RacingCar("2등"))
         )
 
-        finishedRacingCarGame.racing(CarStopNumberGenerator())
+        finishedRacingCarGame.racing { 0 }
 
         finishedRacingCarGame.winners() shouldBe "1등, 2등"
     }
