+++ src/test/kotlin/racingcar/RandomGeneratorTest.kt
@@ -2,10 +2,11 @@ package racingcar
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.ints.shouldBeInRange
-import racingcar.util.RandomGenerator
+import racingcar.domain.NumberGenerator
+import kotlin.random.Random
 
 class RandomGeneratorTest : StringSpec({
-    val randomNumberGenerator = RandomGenerator()
+    val randomNumberGenerator = NumberGenerator { Random.nextInt(0, 10) }
     "랜덤 값 생성 범위는 0 ~ 9 " {
         randomNumberGenerator.generateNumber() shouldBeInRange 0..9
     }
